Minor changes to iterative schemes, tests
InMap is now a function that facilitates the creation of these image/transformed functions as detailed in the last push. (Considering changing the name)

I0 and I1 now take a function handle integrand instead of an InMap instance.

Confirmed over like 2 days that RK4 and IE implementations of I0.I1 are the correct orders of convergence. (Except not really i wouldnt trust that result with my life)