stuff with surface
fixed a thing, actually implemented I1 and geodesicJacobiAB

TODO: apparently matlab supports higher dimensions of vectorized arrays. it could be worth converting geodesicJacobiAB and geodesic so that the output array keeps the shape of the input (rn they're reshaped at the start of the function call). if the plotting functions dont support this, just reshape within the plotting method